+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * Establish target TLS through HTTPS CONNECT proxies #1183
   * Strip brackets from IPv6-literal hosts before TLS SNI/ServerName #1171
   * RequestExt::middleware_config for conf inside middleware #1169